+++ src/color-picker/_example/color-mode.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ColorPicker, Space } from 'tdesign-react';
 
 export default function ColorMode() {
-  const color1 = '#0052d9';
+  const color1 = 'rgba(0, 82, 217, 1)';
   const color2 = '#0052d9';
   const color3 = 'linear-gradient(45deg, #4facfe 0%, #00f2fe 100%)';
 
@@ -14,7 +14,7 @@ export default function ColorMode() {
       </div>
       <div>
         <h5 style={{ marginBottom: 10, fontWeight: 'normal' }}>仅单色模式</h5>
-        <ColorPicker defaultValue={color2} format="CSS" colorModes={['monochrome']} />
+        <ColorPicker defaultValue={color2} format="HEX" colorModes={['monochrome']} />
       </div>
       <div style={{ width: '100%' }}></div>
       <div>