+++ src/color-picker/_example/enable-alpha.jsx
@@ -2,11 +2,11 @@ import React, { useState } from 'react';
 import { ColorPickerPanel } from 'tdesign-react';
 
 export default function EnableAlpha() {
-  const [value, setValue] = useState('#0052d9');
+  const [value, setValue] = useState('#0052D9CC');
 
   const handleChange = (value) => {
     setValue(value);
   };
 
-  return <ColorPickerPanel enableAlpha value={value} format="RGBA" onChange={handleChange} />;
+  return <ColorPickerPanel enableAlpha value={value} format="HEX" onChange={handleChange} />;
 }